+++ tests/Instances/GHC/Event.hs
@@ -1,4 +1,5 @@
-{-# LANGUAGE CPP #-}
+{-# LANGUAGE CPP                #-}
+{-# LANGUAGE StandaloneDeriving #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 
 {-|
@@ -21,8 +22,20 @@ import Prelude.Compat
 
 import Test.QuickCheck (Arbitrary(..), oneof)
 
+# if MIN_VERSION_base(4,8,1)
+import GHC.Event (Lifetime(..))
+import Test.QuickCheck (arbitraryBoundedEnum)
+# endif
+
 instance Arbitrary Event where
     arbitrary = oneof $ map pure [evtRead, evtWrite]
 
 -- TODO: instance Arbitrary FdKey
+
+# if MIN_VERSION_base(4,8,1)
+deriving instance Bounded Lifetime
+deriving instance Enum Lifetime
+instance Arbitrary Lifetime where
+    arbitrary = arbitraryBoundedEnum
+# endif
 #endif